requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asyncio==3.4.3
attrs==24.3.0
audioop-lts==0.2.1
backoff==1.11.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
spotipy==2.25.0
urllib3==2.3.0
yarl==1.18.3
yt-dlp==2025.1.15
